requirements.txt
Flask==2.2.3
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.4.0
tensorflow==2.12.0
tensorflow_intel==2.11.0
